3d_vision/vggt/requirements.txt
einops==0.8.1
gsplat==1.5.3
h5py==3.14.0
imageio==2.37.0
ipdb==0.13.13
lpips==0.1.4
matplotlib==3.10.6
numpy==1.24.0
open3d==0.18.0
opencv_python==4.11.0.86
Pillow==11.3.0
pillow_heif==1.1.0
plyfile==1.1.2
psutil==7.0.0
quaternion==3.5.2.post4
Requests==2.32.5
scikit_learn==1.7.1
scipy==1.15.3
setuptools==80.9.0
tifffile==2025.6.11
torchvision==0.22.1
tqdm==4.67.1
trimesh==4.8.1
viser==1.0.6
gradio==5.44.1
onnxruntime==1.22.1